Prevent Akismet from modifying a comment during checking.
`preprocess_comment` is actually a filter, not an action, so ensure that the data we're returning from it hasn't been modified in any undesired ways.

The wp_filter_comment() call, which was being used to ensure the data we check later matches the data that WordPress asks us to check, was being run on the comment data prematurely, and since it was being returned to the `preprocess_comment` filter, that premature processing was sticking.

See https://github.com/Automattic/jetpack/issues/1365


git-svn-id: https://plugins.svn.wordpress.org/akismet@1076927 b8457f37-d9ea-0310-8a92-e5e31aec5664