Fix handling of comment IDs larger than the largest signed 64-bit integer.
The comment_ID field of the comments table is an unsigned BIGINT, so it can have values as large as twice the largest signed 64-bit integer. PHP's intval() and $wpdb::prepare fail on any number larger than the largest signed integer, so we need to prepare the values as strings.


git-svn-id: https://plugins.svn.wordpress.org/akismet@1481233 b8457f37-d9ea-0310-8a92-e5e31aec5664