Show the default widget title when options are not set.
When the widget is first loaded, it doesn't show the title even if the title is set in the form. The problem was an empty $instance array that is passed to widget() on the first load.

This patch makes the widget() to use the default title when there's no title in the $instance array.

props gonom9


git-svn-id: https://plugins.svn.wordpress.org/akismet@1491264 b8457f37-d9ea-0310-8a92-e5e31aec5664