Use the new helper method to check for a predefined key instead of relying solely on WPCOM_API_KEY.
Sites can now dynamically emulate WPCOM_API_KEY by returning true for the akismet_predefined_key filter and returning their chosen key in the akismet_get_api_key filter. The benefit this has over just setting the wordpress_api_key option is that the user is unaware that the key can be configured and is unable to change it. This makes sense for some hosting environments that might not want to set WPCOM_API_KEY but do want to hide this part of Akismet's configuration.



git-svn-id: https://plugins.svn.wordpress.org/akismet@1691611 b8457f37-d9ea-0310-8a92-e5e31aec5664