Disable Akismet debugging by default unless AKISMET_DEBUG is enabled in wp-config.php.
Initially, we thought that we would have to ask users for this information much more often than we eventually did, which is why it was on by default if WP's debugging was enabled. But since it's turned out to be used less seldom than we expected, I think that requiring a new constant and asking users to enable it is ok, since they would have had to enable WP_DEBUG and WP_DEBUG_LOG too anyway.


git-svn-id: https://plugins.svn.wordpress.org/akismet@1697234 b8457f37-d9ea-0310-8a92-e5e31aec5664