Address an issue that can cause the order of the response values for Jetpack_IXR_ClientMulticall to not match the order of the requests.
The real bug (sort of) is in Jetpack_IXR_ClientMulticall::sort_calls(). It reorders calls (using usort()) to ensure that jetpack.syncContent always gets run first, but it considers all other entries in the array equal.

The manual page for usort() says this: If two members compare as equal, their relative order in the sorted array is undefined. 

Because of this, we can’t rely on calls added by Jetpack_IXR_ClientMulticall::addCall() to ever be in the same order that they were added (or even to be in the same unexpected order). This is less than ideal, because there doesn’t appear to be a way to match response values with the method to which they correspond. For example, if we were calling two methods that could each return an integer between 1 and 100, we would never be able to tell which response went with which call.

If Jetpack must reorder calls, it would be nice if it only moved any jetpack.syncContent calls to the top of the queue and left the order of the rest of the calls alone so that if jetpack.syncContent is not being called, the response order can be trusted.

But since we can’t guarantee that any users running Jetpack and Akismet together will have a version of Jetpack that does this, I’ll write a patch to detect which response value is supposed to be the API key and which should be the user ID.



git-svn-id: https://plugins.svn.wordpress.org/akismet@1756297 b8457f37-d9ea-0310-8a92-e5e31aec5664