Add a REST API method to clear outdated alerts from Akismet.com.
After a key status change, the alert code and message stored on the remote site will not update until the site makes an API request, which is not guaranteed to happen quickly.

With this new endpoint, Akismet.com can ping the site to make a request and retrieve the newest alert code (or more likely, just clear the outdated one).


git-svn-id: https://plugins.svn.wordpress.org/akismet@1965973 b8457f37-d9ea-0310-8a92-e5e31aec5664